refactor: 코드 가독성 리팩토링
