+++ app/src/test/java/nextstep/jwp/view/HtmlViewResolverTest.java
@@ -3,7 +3,6 @@ package nextstep.jwp.view;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.io.IOException;
-import nextstep.jwp.http.response.HtmlView;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
